Crate rustfmt_format_diff

Structs

Opts

Range ğŸ”’

Enums

FormatDiffError ğŸ”’

Constants

DEFAULT_PATTERN ğŸ”’

The default pattern of files to format.

Functions

main ğŸ”’

run ğŸ”’

run_rustfmt ğŸ”’

scan_diff ğŸ”’

Scans a diff from from, and returns the set of files found, and the ranges in those files.